refactor(cli): help
