+++ packages/astro/src/cli/domain/help-payload.ts
