+++ packages/astro/src/cli/infra/build-time-astro-version-provider.ts
